+++ pyobfus_pro/__init__.py
@@ -49,6 +49,7 @@ from .license_binding import (
     LicenseExpired,
     bind_device_key,
     current_machine_id,
+    default_counter_path,
     expire_check,
     period_check,
 )
@@ -126,6 +127,7 @@ __all__ = [
     "_verify_seal",
     "bind_device_key",
     "current_machine_id",
+    "default_counter_path",
     "derive_layer_key",
     "expire_check",
     "forensic_seed",